+++ StockEdSim.Api/Model/Transaction.cs
@@ -6,8 +6,8 @@
         public Guid StudentId { get; set; }
         public virtual ApplicationUser Student { get; set; }
         public string StockSymbol { get; set; }
-        public double Amount { get; set; }  
-        public double PriceAtTransaction { get; set; } 
+        public decimal Amount { get; set; }  
+        public decimal PriceAtTransaction { get; set; } 
         public DateTime TransactionDate { get; set; }
     }
